rolling this all back, i have cooked my goose
