+++ packages/cli/package.json
@@ -51,7 +51,6 @@
     "dotenv": "^8.2.0",
     "execa": "^4.0.0",
     "fs-extra": "^8.1.0",
-    "hasha": "^5.2.2",
     "jju": "^1.4.0",
     "json-keys-sort": "^2.0.0",
     "pkg-up": "^3.1.0",
